chore: move telemetrics into new folder structure
